Files for the Super 8 Camera
There are different versions of the Stl files. they are all very crude to just make it work, finding the perfect spot for the camera is based on trial and error. Good luck
this needs a major redesign, give it a shot and submit a pull request if you made a better version!